app/requirements.txt
kartverket_tide_api
rpi-ws281x
rpi-ws281x-mock
watchdog
bluezero
